Refactor: split VMDetails to separate file
