+++ frontend/packages/kubevirt-plugin/src/components/vms/vm-details.tsx
@@ -1,32 +1,24 @@
 import * as React from 'react';
-import * as _ from 'lodash-es';
+import * as _ from 'lodash';
 
 import { getResource, getServicesForVm } from 'kubevirt-web-ui-components';
 
 import {
-  navFactory,
   Firehose,
   StatusBox,
   ScrollToTopOnMount,
   SectionHeading,
 } from '@console/internal/components/utils';
 
-import { PodKind, K8sResourceKindReference } from '@console/internal/module/k8s';
+import { PodKind } from '@console/internal/module/k8s';
 import { PodModel, ServiceModel } from '@console/internal/models';
-import { DetailsPage } from '@console/internal/components/factory';
 
 import { ServicesList } from '@console/internal/components/service';
 import { VMKind, VMIKind } from './types';
 import { VirtualMachineInstanceModel, VirtualMachineInstanceMigrationModel } from '../../models';
 import { VMResourceSummary, VMDetailsList } from './vm-resource';
 
-// import { VmEvents } from './vm-events';
-// import VmConsolesConnected from '../vmconsoles';
-// import { Nic } from '../nic';
-// import { Disk } from '../disk';
-// import { menuActions } from './menu-actions';
-
-const VMDetailsFirehose = ({ obj: vm }: { obj: VMKind }) => {
+export const VMDetailsFirehose = ({ obj: vm }: { obj: VMKind }) => {
   const { name, namespace } = vm.metadata;
 
   const vmiRes = getResource(VirtualMachineInstanceModel, {
@@ -84,48 +76,6 @@ const VMDetails = (props: VMDetailsProps) => {
       </div>
     </StatusBox>
   );
-};
-
-export const VirtualMachinesDetailsPage = (props: VirtualMachinesDetailsPageProps) => {
-  /* TODO(mlibra): pages will be transferred one by one in follow-ups
-  const consolePage = {
-    href: 'consoles',
-    name: 'Consoles',
-    component: VmConsolesConnected,
-  };
-
-  const nicsPage = {
-    href: 'nics',
-    name: 'Network Interfaces',
-    component: VmNic,
-  };
-
-  const disksPage = {
-    href: 'disks',
-    name: 'Disks',
-    component: VmDisk,
-  };
-  */
-
-  const pages = [
-    navFactory.details(VMDetailsFirehose),
-    navFactory.editYaml(),
-    // consolePage,
-    // navFactory.events(VmEvents),
-    // nicsPage,
-    // disksPage,
-  ];
-
-  const menuActions = undefined; // TODO(mlibra): menuActions
-
-  return <DetailsPage {...props} menuActions={menuActions} pages={pages} />;
-};
-
-type VirtualMachinesDetailsPageProps = {
-  name: string;
-  namespace: string;
-  kind: K8sResourceKindReference;
-  match: any;
 };
 
 type VMDetailsProps = {
